requirements.txt
flask;
pyserial;
python-dotenv